Delete 02_what-does-a-block-do.md
# What does a block do?

When you block a member, they won't be able to see any part of your profile. John has been trying to make this feature for mother-in-laws but no luck so far :-)

Blocks also modifies past exchanges between you and the other person. That means it archives any private messages. It deletes any wall posts, etc. If you were friends, the block eliminates that. It eliminates any relationship statuses between the two of you.

Blocks won't prevent people seeing your comments in groups It blocks the items on your profile.

If the blocked person makes an alternate profile to contact you around that block, that's called a "sock puppet". Notify the FetLife Caretakers, and they'll investigate. If discovered, the sock puppet is deleted and the member warned.